% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_help.R
\name{display_help}
\alias{display_help}
\title{Display the help for function.}
\usage{
display_help(function_name, proj_dir)
}
\arguments{
\item{funtion_name}{The name of the function, str}
}
\description{
Takes the help file '*.Rd' for a function and displays it inline for rmd.
}
\examples{
display_help("make_apa_str")
display_help("display_help")
}
