% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_apa_str.R
\name{make_apa_str}
\alias{make_apa_str}
\title{Make APA String}
\usage{
make_apa_str(out)
}
\arguments{
\item{out}{The output of the test}
}
\value{
APAstr or named list of APAstr if multiple hypothesis tests
}
\description{
Given the output of the tests from Psych 522/524, returns the APA Formatted String.
}
\note{
In the case of an interaction in ANOVA: Reference this as `Factor1xFactor2` and not `Factor1:Factor2`, which is shown in the ANOVA table.
}
\examples{
### T-TEST
data = iris[iris$Species != "virginica",]
(out = t.test(Sepal.Length ~ Species, data = data, var.equal=TRUE))
(make_apa_str(out))

### Chi-Square Test
(tab = table(mtcars$cyl))
(out = chisq.test(tab))
(make_apa_str(out))

### 1 Factor ANOVA
(out = anova(lm(Sepal.Length ~ Species, data=iris)))
make_apa_str(out)
}
