% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_apa_str.R
\name{make_apa_str}
\alias{make_apa_str}
\title{Make APA String}
\usage{
make_apa_str(out)
}
\arguments{
\item{out}{The output of the test}
}
\value{
APAstr or named list of APAstr if multiple hypothesis tests
}
\description{
Given the output of the tests from Psych 522/4,
returns the APA Formatted String.
}
\examples{
out = anova(lm(Sepal.Length ~ Species, data=iris))
make_apa_str(out)
}
